Added support for specifying output path on command line, added timing support, fixed bugs in weight recording and plotting

+++ mad_2007/plot_weights.py
@@ -1,7 +1,7 @@
 import matplotlib.pyplot as plt
 import numpy as np
 
-weights = np.fromfile("weights.bin", dtype=float)
+weights = np.fromfile("weights.bin", dtype=np.float32)
 
 fig, axis = plt.subplots()
 axis.hist(weights, 30, density=True)
